Preserve scene, texture, EFB and audio pointers on ARM64

+++ pc/src/pc_nes_fixnes.c
@@ -91,7 +91,7 @@ extern bool ppuMapper5;
 extern uint8_t inValReads[8];
 
 /* Audio output to game engine */
-extern void AIInitDMA(uint32_t addr, uint32_t size);
+extern void AIInitDMA(uintptr_t addr, uint32_t size);
 
 /* ======================================================================
  * GL rendering state
@@ -219,7 +219,7 @@ static void fixnes_capture_audio(void) {
     }
 
     if (fixnes_audio_pos > 0) {
-        AIInitDMA((uint32_t)(uintptr_t)fixnes_audio_buf, fixnes_audio_pos * 4);
+        AIInitDMA((uintptr_t)fixnes_audio_buf, fixnes_audio_pos * 4);
     }
 }
