+++ src/static/jaudio_NES/internal/aictrl.c
@@ -60,7 +60,7 @@ extern void Jac_Init(void) {
     }
 
     AIInit(nullptr);
-    AIInitDMA((u32)dac[2], DAC_SIZE * 2);
+    AIInitDMA((uintptr_t)dac[2], DAC_SIZE * 2);
 }
 
 static void MixMonoTrack(s16* track, s32 nSamples, MixCallback callback) {
@@ -282,7 +282,7 @@ extern void Jac_UpdateDAC(void) {
     }
 
     if (use_rsp_madep != nullptr) {
-        AIInitDMA((u32)use_rsp_madep, DAC_SIZE * 2);
+        AIInitDMA((uintptr_t)use_rsp_madep, DAC_SIZE * 2);
         use_rsp_madep = nullptr;
     } else {
         UNIVERSAL_DACCOUNTER++;
